switch to Wii U Pro Controller
every single problem I had with the moga 2 controller just... completely
went away. this is a much better controller.

+++ src/hw/input/js.h
@@ -23,7 +23,7 @@ class JoystickInput {
  private:
   int fd_;
 
-  uint16_t buttons_;
+  uint32_t buttons_;
   int16_t axes_[8];
 };
